test(card-events): move the bearer tests into the mirrored file
Same PS-204 §2 orphan-test failure as the sibling PR: a test file must mirror a
src module, and `test__card_event_bearer_env_alias.py` was named for the BUG
rather than the module, so it mirrored nothing.

    [E] [PS-204 §2 orphan-test-file] ... no matching src file (orphan test)

The five bearer cases now live in `_listen/test__card_event_delivery.py`, next
to the other tests for the same module — a file that already carried
"Mirrors src/scitex_agent_container/_listen/_card_event_delivery.py (PS-204 §2)"
in its own docstring. The convention was written down, in the very file the
tests belonged in, and I did not follow it.

Local runs did not catch this because I ran only the new file; the project
audit runs INSIDE the CI pytest invocation, so a single-file run skips it
entirely. Running one test file is not running the suite — the same lesson the
sibling PR records.
